Create Form View and Item View

+++ src/containers/ItemView.js
@@ -4,29 +4,22 @@ import { browserHistory } from 'react-router';
 
 import SubMenu from '../components/common/SubMenu';
 import BusyIcon from '../components/common/BusyIcon';
-import EventForm from '../components/item/EventForm';
 import EventItem from '../components/item/EventItem';
 
-import { loadTags, unloadTags } from '../actions/tags';
-import { loadEvent, unloadEvent, createEvent, updateEvent, deleteEvent } from '../actions/event';
+import { loadEvent } from '../actions/event';
 import { unloadCachedData } from '../actions/events';
-import { resetFilters } from '../actions/common';
+
+import '../styles/item-view.css';
 
 class ItemView extends Component {
   constructor(props, context) {
     super(props, context);
-    this.onSubmit = this.onSubmit.bind(this);
     this.onModify = this.onModify.bind(this);
-    this.onCancel = this.onCancel.bind(this);
     this.onDelete = this.onDelete.bind(this);
 
     this.state = {
-      // Show form or overview
-      editing: false,
-      // Show or hide scrim
       changing: false,
-      item: { ...props.item
-      }
+      item: null
     };
   }
 
@@ -34,19 +27,10 @@ class ItemView extends Component {
     if (this.props.params.id) {
       // Load the item
       this.props.dispatch(loadEvent(this.props.params.id));
-    } else {
-      this.beginEdit();
     }
   }
 
   componentWillReceiveProps(newProps) {
-    if (this.props.params.id && !newProps.params.id) {
-      // Add button clicked on this component
-      // Clear the item and begin to edit mode
-      this.props.dispatch(unloadEvent());
-      this.beginEdit();
-    }
-
     if (this.props.item != newProps.item) {
       this.setState({
         item: {
@@ -56,54 +40,9 @@ class ItemView extends Component {
     }
   }
 
-  componentWillUnmount() {
-    this.props.dispatch(unloadEvent());
-    return true;
-  }
-
-  // create of update item
-  onSubmit(newItem) {
-    // clear the annual cache
-    const onComplete = (id) => {
-      // clear the cache
-      this.props.dispatch(unloadCachedData());
-      this.props.dispatch(unloadTags());
-      this.props.dispatch(resetFilters());
-
-      this.setState({
-        editing: false,
-        changing: false
-      });
-
-      if (id) {
-        // When new item created, a id of new record will be called back
-        // Redirect to new created item
-        browserHistory.push(`/event/${id}`);
-      }
-    };
-
-    this.setState({
-      item: newItem,
-      changing: true
-    });
-
-    if (newItem.id == null) {
-      this.props.dispatch(createEvent(newItem, onComplete));
-    } else {
-      this.props.dispatch(updateEvent(newItem, onComplete));
-    }
-  }
-
-  // switch to the edit mode
-  onModify() {
-    this.beginEdit();
-  }
-
-  // cancel the edit mode
-  onCancel() {
-    this.setState({
-      editing: false
-    });
+  // redirect to the edit view
+  onModify(id) {
+    browserHistory.push(`/edit/${id}`);
   }
 
   // delete item
@@ -116,88 +55,39 @@ class ItemView extends Component {
       browserHistory.push('/');
     };
 
-    this.setState({
-      changing: true
-    });
+    if (window.confirm('Are you sure to delete?')) {
+      this.setState({
+        changing: true
+      });
 
-    this.props.dispatch(deleteEvent(id, onComplete));
-  }
-
-  beginEdit() {
-    if (this.props.tags.length <= 0) {
-      // get all tags for the tag-radio
-      this.props.dispatch(loadTags());
+      this.props.dispatch(deleteEvent(id, onComplete));
     }
-
-    this.setState({
-      editing: true
-    });
   }
 
   render() {
-    const id = this.props.params.id;
     const {
-      activeYear,
-      activeMonth,
-      activeDate,
       previousView
     } = this.props;
     const item = this.state.item;
 
-    let title;
-    let bodyDOM;
-
-    if (id && item.id == null) {
-      // loading
-      title = 'Loading...';
-      bodyDOM = (
-        <div style={{ padding: '20px 10px' }}>
-          <BusyIcon />
-        </div>
-      );
-    } else if (this.state.editing) {
-      if (id) {
-        // editing
-        title = 'Modify Event';
-      } else {
-        // creating
-        title = 'New Event';
-
-        // initial year, month and date
-        const date = new Date();
-        item.year = activeYear === 0 ? date.getFullYear() : activeYear;
-        item.month = activeMonth === 0 ? date.getMonth() + 1 : activeMonth;
-        item.date = activeDate === 0 ? date.getDate() : activeDate;
-      }
-      bodyDOM = (
-        <EventForm
-          item={item}
-          tags={this.props.tags}
-          changing={this.state.changing}
-          onSubmit={this.onSubmit}
-          onCancel={this.onCancel}
-        />
-      );
-    } else {
-      title = `${item.year}-${item.month}-${item.date}`;
-      bodyDOM = (
-        <EventItem
-          item={item}
-          changing={this.state.changing}
-          onModify={this.onModify}
-          onDelete={this.onDelete}
-        />
-      );
+    if (!item) {
+      return <div style={{ margin: '15px 10px' }}><BusyIcon /></div>;
     }
 
     return (
       <div>
         <SubMenu previousView={previousView}>
           <div className="right">
-            {title}
+            {`${item.year}-${item.month}-${item.date}`}
           </div>
         </SubMenu>
-        {bodyDOM}
+        <div className="item-view">
+          <EventItem item={item} />
+          <div className="tools">
+            <button type="button" className="button" onClick={() => this.onModify(item.id)}>Modify</button>
+            <button type="button" className="button" onClick={() => this.onDelete(item.id)}>Delete</button>
+          </div>
+        </div>
       </div>
     );
   }
@@ -205,21 +95,12 @@ class ItemView extends Component {
 
 ItemView.propTypes = {
   item: PropTypes.object,
-  tags: PropTypes.array,
-  activeYear: PropTypes.number,
-  activeMonth: PropTypes.number,
-  activeDate: PropTypes.number,
   previousView: PropTypes.object,
-  params: PropTypes.object,
   dispatch: PropTypes.func
 };
 
 const mapStateToProps = (state) => ({
   item: state.event.item,
-  tags: state.tags.list,
-  activeYear: state.events.activeYear,
-  activeMonth: state.events.activeMonth,
-  activeDate: state.events.activeDate,
   previousView: state.common.previousView
 });
